Renamed logger to evaluation_runner to make its purpose more clear as the entry point. Began full implementation.
